+++ src/protocols/secure_channel/BUILD.gn
@@ -11,6 +11,7 @@ static_library("secure_channel") {
     "PASESession.cpp",
     "PASESession.h",
     "RendezvousParameters.h",
+    "SessionEstablishmentDelegate.h",
     "SessionEstablishmentExchangeDispatch.cpp",
     "SessionEstablishmentExchangeDispatch.h",
     "StatusReport.cpp",
