+++ src/protocols/secure_channel/SessionEstablishmentDelegate.h
